execution_profile: Moved defaults mod here from Session
As argued, it is only used by execution profiles.
