+++ scylla/src/transport/session.rs
@@ -62,56 +62,6 @@ use crate::authentication::AuthenticatorProvider;
 #[cfg(feature = "ssl")]
 use openssl::ssl::SslContext;
 
-/// Stores default (somehow recommended) configuration options for Session.
-pub mod defaults {
-    use scylla_cql::frame::types::SerialConsistency;
-    use scylla_cql::Consistency;
-    use std::sync::Arc;
-    use std::time::Duration;
-
-    use crate::load_balancing::{LoadBalancingPolicy, RoundRobinPolicy, TokenAwarePolicy};
-    use crate::retry_policy::{DefaultRetryPolicy, RetryPolicy};
-    use crate::speculative_execution::SpeculativeExecutionPolicy;
-    use crate::transport::execution_profile::ExecutionProfileInner;
-
-    pub fn consistency() -> Consistency {
-        Consistency::LocalQuorum
-    }
-
-    pub fn serial_consistency() -> Option<SerialConsistency> {
-        None
-    }
-
-    pub fn request_timeout() -> Option<Duration> {
-        Some(Duration::from_secs(30))
-    }
-
-    pub fn load_balancing_policy() -> Arc<dyn LoadBalancingPolicy> {
-        Arc::new(TokenAwarePolicy::new(Box::new(RoundRobinPolicy::new())))
-    }
-
-    pub fn retry_policy() -> Box<dyn RetryPolicy> {
-        Box::new(DefaultRetryPolicy::new())
-    }
-
-    pub fn speculative_execution_policy() -> Option<Arc<dyn SpeculativeExecutionPolicy>> {
-        None
-    }
-
-    impl Default for ExecutionProfileInner {
-        fn default() -> Self {
-            Self {
-                request_timeout: request_timeout(),
-                consistency: consistency(),
-                serial_consistency: serial_consistency(),
-                load_balancing_policy: load_balancing_policy(),
-                retry_policy: retry_policy(),
-                speculative_execution_policy: speculative_execution_policy(),
-            }
-        }
-    }
-}
-
 #[derive(Debug, Copy, Clone)]
 pub enum TranslationError {
     NoRuleForAddress,
